requirements.txt
absl-py==2.1.0
accelerate==1.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.82
dataclasses-json==0.6.7
distro==1.9.0
etils==1.5.2
exceptiongroup==1.2.2
filelock==3.16.1
flax==0.7.0
frozenlist==1.5.0
fsspec==2024.10.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-text-splitters==0.3.2
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
milvus==2.3.9
milvus-lite==2.4.10
ml_dtypes==0.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
ollama==0.4.4
openai==1.57.2
opt_einsum==3.4.0
optax==0.1.4
orbax-checkpoint==0.5.16
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pymilvus==2.5.0
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.4.5
scipy==1.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
tensorstore==0.1.69
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
yarl==1.18.3
zipp==3.21.0
